Module assembly_xml::database

The XML <database> format

Before the FDB file format was created, older versions of the client used an XML file to store the client database. This was also used for LUPs to add their data independently of the rest of the game.

Structs

Column

The information on a column

Row

A row of the database

UnknownValueType

Unknown value type ‘{0}’

Enums

ValueType

The value types for the database

Functions

expect_column_or_end_columns

Expects an empty <column …/> tag or a closing </columns> tag

expect_columns

Expects an opening <columns> tag

expect_database

Expects an opening <database>

expect_row_or_end_rows

Expects an empty <row …/> tag or a closing </rows> tag

expect_rows

Expects an opening <rows> tag

expect_table

Expects an opening <table> tag or a closing </database> tag